% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangevoting.R
\name{majority}
\alias{majority}
\title{Majority system calculator}
\usage{
majority(elected, counter)
}
\arguments{
\item{elected}{"Array of representatives elected in previously analyzed costituencies"}

\item{counter}{"Counter of total parliament seats for each party in previously analyzed costituencies"}
}
\value{
The updated array of elected representatives

Updated counter of total parliament seats
}
\description{
Reads a csv file containing votes and computes the winner of that costituency in a majority system
}
